+++ src/main/java/org/odessajavaclub/user/application/service/GetUsersService.java
@@ -1,9 +1,7 @@
 package org.odessajavaclub.user.application.service;
 
 import lombok.RequiredArgsConstructor;
-import org.odessajavaclub.user.application.port.in.GetUserQuery;
 import org.odessajavaclub.user.application.port.in.GetUsersQuery;
-import org.odessajavaclub.user.application.port.out.LoadUserPort;
 import org.odessajavaclub.user.application.port.out.LoadUsersPort;
 import org.odessajavaclub.user.domain.User;
 import org.springframework.stereotype.Service;
@@ -13,9 +11,7 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
-class GetUsersService implements GetUsersQuery, GetUserQuery {
-
-    private final LoadUserPort loadUserPort;
+class GetUsersService implements GetUsersQuery {
 
     private final LoadUsersPort loadUsersPort;
 
@@ -26,6 +22,6 @@ class GetUsersService implements GetUsersQuery, GetUserQuery {
 
     @Override
     public Optional<User> getUser(UserQuery userQuery) {
-        return Optional.ofNullable(loadUserPort.loadUser(userQuery.getUserId()));
+        return Optional.ofNullable(loadUsersPort.loadUser(userQuery.getUserId()));
     }
 }
