+++ src/main/java/org/odessajavaclub/user/application/service/UpdateUserService.java
@@ -2,7 +2,7 @@ package org.odessajavaclub.user.application.service;
 
 import lombok.RequiredArgsConstructor;
 import org.odessajavaclub.user.application.port.in.UpdateUserUseCase;
-import org.odessajavaclub.user.application.port.out.LoadUserPort;
+import org.odessajavaclub.user.application.port.out.LoadUsersPort;
 import org.odessajavaclub.user.application.port.out.UpdateUserPort;
 import org.odessajavaclub.user.domain.User;
 import org.springframework.stereotype.Service;
@@ -14,7 +14,7 @@ import java.util.Optional;
 @RequiredArgsConstructor
 class UpdateUserService implements UpdateUserUseCase {
 
-    private final LoadUserPort loadUserPort;
+    private final LoadUsersPort loadUserPort;
 
     private final UpdateUserPort updateUserPort;
